requirements.txt
gradio
transformers
torch
requests
python-dotenv